Update mazeRunner.py
The ScreenShots of the example run under assignments does not line up. This update allows students to get the same feedback via the console, just as it should be if the students are going based off of the images. I recommend removing the pseudocode because the students know what to do based on what was given under assignments. They needed "help" because grade scope wouldnt accept their solution based on what was given by assignment